c4ReplicatorImpl now supports DatabasePool
